Separate broker host platform bindings
Move the Linux-userland host control-channel endpoints into a dedicated platform crate while preserving the existing local endpoint and shared control-ring implementation.


Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
