+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "litebox_broker_core",
     "litebox_broker_protocol",
     "litebox_broker_host",
+    "litebox_broker_platform_linux_userland",
     "litebox_broker_transport",
     "litebox_broker_userland",
     "litebox_common_linux",
@@ -41,6 +42,7 @@ default-members = [
     "litebox_broker_core",
     "litebox_broker_protocol",
     "litebox_broker_host",
+    "litebox_broker_platform_linux_userland",
     "litebox_broker_transport",
     "litebox_broker_userland",
     "litebox_common_linux",
