+++ litebox_runner_linux_userland/Cargo.toml
@@ -25,6 +25,7 @@ walkdir = "2.0"
 glob = "0.3"
 litebox_broker_core = { version = "0.1.0", path = "../litebox_broker_core" }
 litebox_broker_host = { version = "0.1.0", path = "../litebox_broker_host" }
+litebox_broker_platform_linux_userland = { version = "0.1.0", path = "../litebox_broker_platform_linux_userland" }
 
 [features]
 lock_tracing = ["litebox/lock_tracing"]
